+++ examples/src/WebGL2Samples/glsl_centroid.ts
@@ -164,7 +164,7 @@ for (let i = 0; i < VIEWPORTS.MAX; ++i)
         descriptor: framebuffers[i],
         renderPassObjects: [{
             pipeline: programs[i],
-            bindingResources: { MVP: IDENTITY },
+            bindingResources: { MVP: IDENTITY as Float32Array },
             vertices: vertexArrays[i].vertices,
             draw: { __type__: "DrawVertex", vertexCount },
         }]
@@ -196,7 +196,7 @@ for (let i = 0; i < VIEWPORTS.MAX; ++i)
             ...ro,
             viewport: viewport[i],
             bindingResources: {
-                MVP: mvp,
+                MVP: mvp as Float32Array,
                 diffuse: { texture: textures[i], sampler: samplers[i] },
             },
             draw: { __type__: "DrawVertex", vertexCount: 6 },
