+++ examples/src/webgl-examples/sample4.ts
@@ -92,8 +92,8 @@ function main()
 
     const { projectionMatrix, modelViewMatrix } = drawScene(canvas, deltaTime);
 
-    reactive(renderObject.bindingResources).uProjectionMatrix = projectionMatrix;
-    reactive(renderObject.bindingResources).uModelViewMatrix = modelViewMatrix;
+    reactive(renderObject.bindingResources).uProjectionMatrix = projectionMatrix as Float32Array;
+    reactive(renderObject.bindingResources).uModelViewMatrix = modelViewMatrix as Float32Array;
 
     webgl.submit({ commandEncoders: [{ passEncoders: [renderPass] }] });
